refactor: extract match routes from resumes_search.ts into resumes_match.ts
Split the 2211-line resumes_search.ts into two focused route modules:
- resumes_search.ts (1156 lines): list samples, keyword expansion, search
- resumes_match.ts (1095 lines): match, match-stream, matches, match-runs, clear

Match-only helpers (stripFrontMatter, extractSection, buildMatchQueryMetadata,
buildRuleMatchResponseEntry, mapStoredMatch, mapStoredMatchRun, toMatchMode,
toTopN, computeStats) moved to resumes_match.ts. Shared helpers
(scorePreparedCandidates, prepareSampleCandidates) remain in resumes_search.ts
and are exported for cross-module use.
