+++ apps/api/src/routes/index.ts
@@ -25,6 +25,7 @@ export { default as webVitalsRoutes } from "./web-vitals.js";
 export { default as searchAlertsRoutes } from "./search-alerts.js";
 export { default as resumesAdminRoutes } from "./resumes_admin.js";
 export { default as resumesSearchRoutes } from "./resumes_search.js";
+export { default as resumesMatchRoutes } from "./resumes_match.js";
 export { default as resumesDiagnosticsRoutes } from "./resumes_diagnostics.js";
 export { default as resumesImportRoutes } from "./resumes_import.js";
 export { default as resumesPacketsRoutes } from "./resumes_packets.js";
